[変更]DB定義ファイルを分割
